+++ backend/config/settings/base.py
@@ -332,24 +332,55 @@ warnings.simplefilter("always", DeprecationWarning)
 # ------------------------------------------------------------------------------
 # https://django-storages.readthedocs.io/en/latest/#installation
 
+# Get the storage backend
+STORAGE_BACKEND = env.str("DJANGO_STORAGE_BACKEND", "storages.backends.s3.S3Storage")
+
+# Define storage options based on the selected backend
+storage_options = {}
+
+# S3 Storage options
+if STORAGE_BACKEND == "storages.backends.s3.S3Storage":
+    storage_options = {
+        # Fly.io Tigris
+        # The strange environment variable names are based on what Fly.io provides when setting
+        # up Tigris.
+        "endpoint_url": env.str("AWS_ENDPOINT_URL_S3", None),
+        "region_name": env.str("AWS_REGION", None),
+        "bucket_name": env.str("BUCKET_NAME", None),
+        "default_acl": "private",
+        "querystring_auth": env.bool("AWS_QUERYSTRING_AUTH", True),
+        "querystring_expire": env.int("AWS_QUERYSTRING_EXPIRE", 3600 * 12),
+        "location": env.str("AWS_LOCATION", ""),
+        "custom_domain": env.str("AWS_S3_CUSTOM_DOMAIN", None),
+        "url_protocol": env.str("AWS_S3_URL_PROTOCOL", "https:"),
+        "addressing_style": env.str("AWS_S3_ADDRESSING_STYLE", None),
+    }
+# Azure Storage options
+elif STORAGE_BACKEND == "storages.backends.azure_storage.AzureStorage":
+    storage_options = {
+        "azure_container": env.str("AZURE_CONTAINER", None),
+        "account_name": env.str("AZURE_ACCOUNT_NAME", None),
+        "account_key": env.str("AZURE_ACCOUNT_KEY", None),
+        "custom_domain": env.str("AZURE_CUSTOM_DOMAIN", None),
+        "azure_ssl": env.bool("AZURE_SSL", True),
+        "connection_string": env.str("AZURE_CONNECTION_STRING", None),
+        "token_credential": None,  # This needs to be set programmatically if used
+        "overwrite_files": env.bool("AZURE_OVERWRITE_FILES", True),  # We expect this to be true
+        "location": env.str("AZURE_LOCATION", ""),
+        "cache_control": env.str("AZURE_CACHE_CONTROL", None),
+        "endpoint_suffix": env.str("AZURE_ENDPOINT_SUFFIX", None),
+        "expiration_secs": env.int("AZURE_URL_EXPIRATION_SECS", 3600 * 12),
+    }
+
 STORAGES = {
     "default": {
-        "BACKEND": "storages.backends.s3.S3Storage",
-        "OPTIONS": {
-            # Fly.io Tigris
-            # The strange environment variable names are based on what Fly.io provides when setting
-            # up Tigris.
-            "endpoint_url": env.str("AWS_ENDPOINT_URL_S3", None),
-            "region_name": env.str("AWS_REGION"),
-            "bucket_name": env.str("BUCKET_NAME"),
-            "default_acl": "private",
-            "querystring_auth": env.bool("AWS_QUERYSTRING_AUTH", True),
-            "querystring_expire": env.int("AWS_QUERYSTRING_EXPIRE", 3600 * 12),
-            "location": env.str("AWS_LOCATION", ""),
-            "custom_domain": env.str("AWS_S3_CUSTOM_DOMAIN", None),
-            "url_protocol": env.str("AWS_S3_URL_PROTOCOL", "https:"),
-            "addressing_style": env.str("AWS_S3_ADDRESSING_STYLE", None),
-        },
+        # The storage backend can be configured using the DJANGO_STORAGE_BACKEND environment
+        # variable.
+        # Options are:
+        # - "storages.backends.s3.S3Storage" (default)
+        # - "storages.backends.azure_storage.AzureStorage"
+        "BACKEND": STORAGE_BACKEND,
+        "OPTIONS": storage_options,
     },
     "staticfiles": {"BACKEND": "django.contrib.staticfiles.storage.StaticFilesStorage"},
 }